requirements.txt
beautifulsoup4==4.9.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorlog==4.2.1
cryptography==3.1.1
decorator==4.4.2
idna==2.10
markkk==0.0.12
pkg-resources==0.0.0
pycparser==2.20
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.25.10
